+++ MiscUtils.h
@@ -125,7 +125,7 @@ int MscMD5Authenticate(char const *pszPassword, char const *pszTimeStamp, char c
 char *MscExtractServerTimeStamp(char const *pszResponse, char *pszTimeStamp, int iMaxTimeStamp);
 int MscRootedName(char const *pszHostName);
 int MscCramMD5(char const *pszSecret, char const *pszChallenge, char *pszDigest);
-unsigned long MscHashString(char const *pszBuffer, int iLength,
+unsigned long MscHashString(char const *pszBuffer, size_t iLength,
 			    unsigned long ulHashInit = HASH_INIT_VALUE);
 int MscSplitAddressPort(char const *pszConnSpec, char *pszAddress, int &iPortNo, int iDefPortNo);
 SYS_UINT16 MscReadUint16(void const *pData);
@@ -151,4 +151,3 @@ int MscStringCompareCB(void *pPrivate, HashDatum const *pDatum1,
 		       HashDatum const *pDatum2);
 
 #endif
-
